KTOR-917 Keep some characters in cookie URI encoding according to RFC 3986
